requirements.txt
plyer
python-dotenv
requests
twilio
